backend/pyproject.toml
[project]
name = "dolce-backend"
version = "1.0.0"
description = "Django REST Framework backend for Dolce Fiore e-commerce platform"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15.0",
    "drf-spectacular>=0.27.0",
    "drf-spectacular-sidecar>=2024.5.1",
    "django-cors-headers>=4.3.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

